Network Policy yaml
Create a new NetworkPolicy named NP1 in the namespace NS1. Ensure that the new NetworkPolicy allows Pods in namespace NS2 to connect to port 8080 of Pods in namespace NS1. Further ensure that the new NetworkPolicy NP1 does not allow access to Pods, which don't listen on port 8080 and does not allow access from Pods, which are not in namespace NS2.